# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Maintenance Copilot",
    "description": "Multi-agent industrial assistant powered by LangGraph, SQLite checkpointing, FAISS RRF, and Argon2 RBAC.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/equipment/{machine_id}/alarms": {
      "get": {
        "tags": [
          "Equipment v1"
        ],
        "summary": "Get Equipment Alarms",
        "description": "Retrieve active or historical alarms for a specific machine.",
        "operationId": "get_equipment_alarms_api_v1_equipment__machine_id__alarms_get",
        "parameters": [
          {
            "name": "machine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Machine Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by alarm status: active, acknowledged, cleared",
              "title": "Status"
            },
            "description": "Filter by alarm status: active, acknowledged, cleared"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/alarms": {
      "get": {
        "tags": [
          "Alarms v1"
        ],
        "summary": "List Alarms",
        "description": "List alarms across the plant, newest first.",
        "operationId": "list_alarms_api_v1_alarms_get",
        "parameters": [
          {
            "name": "machine_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Machine Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "active",
                    "acknowledged",
                    "cleared"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alarms/{alarm_id}/acknowledge": {
      "post": {
        "tags": [
          "Alarms v1"
        ],
        "summary": "Acknowledge Alarm",
        "description": "Acknowledge an active alarm. Recorded in the audit log.",
        "operationId": "acknowledge_alarm_api_v1_alarms__alarm_id__acknowledge_post",
        "parameters": [
          {
            "name": "alarm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alarm Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sops": {
      "get": {
        "tags": [
          "SOPs v1"
        ],
        "summary": "List Sops",
        "description": "Search the SOP catalogue.",
        "operationId": "list_sops_api_v1_sops_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "safety",
                    "maintenance",
                    "troubleshooting",
                    "operation"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "asset",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "active",
                    "review",
                    "draft"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "SOPs v1"
        ],
        "summary": "Create Sop",
        "description": "Add a draft SOP (admin only). Rebuild the document index for the copilot to cite it.",
        "operationId": "create_sop_api_v1_sops_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SopCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/history": {
      "get": {
        "tags": [
          "History v1"
        ],
        "summary": "List History",
        "description": "Alarms, diagnostics, work orders, maintenance, approvals, SOP views, and settings changes, newest first.",
        "operationId": "list_history_api_v1_history_get",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive start date (YYYY-MM-DD)",
              "title": "From"
            },
            "description": "Inclusive start date (YYYY-MM-DD)"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exclusive end date (YYYY-MM-DD)",
              "title": "To"
            },
            "description": "Exclusive end date (YYYY-MM-DD)"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "alarm",
                    "diagnostic",
                    "work_order",
                    "maintenance",
                    "approval",
                    "sop",
                    "settings"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "machine_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 50
                },
                {
                  "type": "null"
                }
              ],
              "title": "Machine Id"
            }
          },
          {
            "name": "user",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "User"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/equipment/{machine_id}/history": {
      "get": {
        "tags": [
          "Compatibility"
        ],
        "summary": "Machine History Legacy",
        "operationId": "machine_history_legacy_api_equipment__machine_id__history_get",
        "parameters": [
          {
            "name": "machine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Machine Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SopCreate": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^SOP-[A-Z0-9\\-]{2,30}$",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "maxLength": 150,
            "minLength": 3,
            "title": "Title"
          },
          "category": {
            "type": "string",
            "enum": [
              "safety",
              "maintenance",
              "troubleshooting",
              "operation"
            ],
            "title": "Category"
          },
          "assets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Assets"
          },
          "body": {
            "type": "string",
            "maxLength": 50000,
            "minLength": 20,
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "category",
          "body"
        ],
        "title": "SopCreate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}